requirements.txt
gitdb2==2.0.3
GitPython==2.1.10
smmap2==2.0.3
